% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_data2007.R
\docType{data}
\name{data2007}
\alias{data2007}
\title{Example data from lecture 6}
\format{An object of class \code{data.frame} with 142 rows and 6 columns.}
\usage{
data2007
}
\description{
Example data from lecture 6
}
\references{
\url{http://users.stat.umn.edu/~almquist/3811_examples/gapminder2007ex.csv}
}
\author{
Zack almquist \email{almquist@umn.edu}
}
\keyword{datasets}
