% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ2_Rfunctions.R
\name{func10}
\alias{func10}
\title{standardize matrix}
\usage{
func10(x)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix
}
\description{
standardize matrix
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
func10(a)
}
