% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ2_Rfunctions.R
\name{func11}
\alias{func11}
\title{standardize matrix without loop}
\usage{
func11(x)
}
\arguments{
\item{a}{matrix}
}
\value{
x matrix
}
\description{
standardize matrix
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
func11(a)
}
