% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ1_Rfunctions.R
\name{func6}
\alias{func6}
\title{Highlevel check function}
\usage{
func6(x, y)
}
\arguments{
\item{x}{object}
}
\value{
object
}
\description{
Checks and throws error if not numeric, finit, zero lenth, NA, NAN
}
\examples{
func6(1,Inf)
}
