% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ2_Rfunctions.R
\name{func8}
\alias{func8}
\title{calculate matrix multiplcation}
\usage{
func8(a, x)
}
\arguments{
\item{a}{matrix}

\item{x}{matrix}
}
\value{
matrix
}
\description{
calculates $x^T A^{-1} x$
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
func8(a,x)
}
