% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ2_Rfunctions.R
\name{\%func9\%}
\alias{\%func9\%}
\title{binary operator function}
\usage{
a \%func9\% x
}
\arguments{
\item{a}{matrix}

\item{x}{matrix}
}
\value{
matrix
}
\description{
calculates $x^T A^{-1} x$
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
a \%func9\% x
}
