% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_m.R
\docType{data}
\name{m}
\alias{m}
\title{data from homework 2}
\format{An object of class \code{url} (inherits from \code{connection}) of length 1.}
\usage{
m
}
\description{
data from homework 2
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q2p3.rda}
}
\author{
Charlie Geyer @author Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
