% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsHW&QUIZ2_Rfunctions.R
\name{myapply}
\alias{myapply}
\title{myapply function}
\usage{
myapply(x, MARGIN, FUN, ...)
}
\arguments{
\item{x}{an array}

\item{MARGIN}{a vector giving the subscripts which the function will be applied over.}

\item{FUN}{functions to be applied}

\item{...}{optional arguments to FUN}
}
\value{
array
}
\description{
it is similar with apply function
}
\examples{
fred=matrix(1:6,nrow = 3)
rownames(fred)=c("red","white","blue")
colnames(fred)=c("car","truck")
myapply(fred, 1, mean)
}
