% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JiangZhihangToolsquiz-myplot.R
\name{plotMyData}
\alias{plotMyData}
\title{Wrapper function for ggplot2 for data d}
\usage{
plotMyData(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
plot data using ggplot
}
\examples{
data(d)
plotMyData(d)
}
